requirements.txt
PySide2
pyyaml